feat : inventory system full feature v1 + postman collection

+++ src/main/java/com/project/inventory/service/impl/InventoryServiceImpl.java
@@ -1,10 +1,13 @@
 package com.project.inventory.service.impl;
 
-import com.project.inventory.dto.InventoryDTO;
+import com.project.inventory.dto.request.InventoryRequest;
+import com.project.inventory.dto.request.UpdateInventoryRequest;
 import com.project.inventory.dto.response.InventoryResponse;
 import com.project.inventory.model.Inventory;
 import com.project.inventory.repository.InventoryRepository;
+import com.project.inventory.repository.ItemRepository;
 import com.project.inventory.service.InventoryService;
+import com.project.inventory.service.ItemService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -15,6 +18,10 @@ import org.springframework.stereotype.Service;
 public class InventoryServiceImpl implements InventoryService {
 
     private final InventoryRepository inventoryRepository;
+
+    private final ItemRepository itemRepository;
+
+    private final ItemService itemService;
 
     @Override
     public InventoryResponse get(Long id) {
@@ -34,13 +41,20 @@ public class InventoryServiceImpl implements InventoryService {
     }
 
     @Override
-    public InventoryResponse save(InventoryDTO inventory) {
+    public InventoryResponse save(InventoryRequest inventory) {
+        Integer stock = itemService.getRemainingStock(inventory.getItemId());
+
+        if (inventory.getType().equals("W") && stock < inventory.getQty()) {
+            throw new RuntimeException("Insufficient stock");
+        }
+
         Inventory savedInventory = Inventory.builder()
                 .type(inventory.getType())
-                .item(inventory.getItem())
+                .item(itemRepository.findById(inventory.getItemId()).orElseThrow(()-> new RuntimeException("Item not found")))
                 .qty(inventory.getQty())
                 .build();
         inventoryRepository.save(savedInventory);
+
         return InventoryResponse.builder()
                 .id(savedInventory.getId())
                 .item(savedInventory.getItem())
@@ -50,11 +64,17 @@ public class InventoryServiceImpl implements InventoryService {
     }
 
     @Override
-    public InventoryResponse update(Long id, InventoryDTO inventory) {
+    public InventoryResponse update(Long id, UpdateInventoryRequest inventoryRequest) {
         Inventory inventoryExisting = inventoryRepository.findById(id).orElseThrow(()-> new RuntimeException("Inventory not found"));
 
-        inventoryExisting.setQty(inventory.getQty());
-        inventoryExisting.setType(inventory.getType());
+        Integer currentStock = itemService.getRemainingStock(id) - inventoryExisting.getQty();
+
+        if (inventoryRequest.getType().equals("W") && (currentStock - inventoryRequest.getQty() < 0)) {
+            throw new RuntimeException("Insufficient stock");
+        }
+
+        inventoryExisting.setQty(inventoryRequest.getQty());
+        inventoryExisting.setType(inventoryRequest.getType());
 
         inventoryRepository.save(inventoryExisting);
 
@@ -67,7 +87,14 @@ public class InventoryServiceImpl implements InventoryService {
     }
 
     @Override
-    public void delete(Long id) {
+    public String delete(Long id) {
+
+        Inventory currentInventory = inventoryRepository.findById(id).orElseThrow(()-> new RuntimeException("Inventory not found"));
+        int currentStock = itemService.getRemainingStock(id) - currentInventory.getQty();
+        if (currentInventory.getType().equals("T") && (currentStock < 0)) {
+            throw new RuntimeException("Insufficient stock");
+        }
         inventoryRepository.deleteById(id);
+        return String.format("Inventory with id : %d deleted successfully", id);
     }
 }
